+++ src/Application/Cart/UpdateBillingAddress.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Thinktomorrow\Trader\Application\Cart;
@@ -10,13 +11,20 @@ use Thinktomorrow\Trader\Domain\Model\Order\OrderId;
 class UpdateBillingAddress
 {
     private string $orderId;
+
     private ?string $countryId;
+
     private ?string $line1;
+
     private ?string $line2;
+
     private ?string $postalCode;
+
     private ?string $city;
 
-    public function __construct(string $orderId, ?string $countryId = null, ?string $line1 = null, ?string $line2 = null, ?string $postalCode = null, ?string $city = null)
+    private array $data;
+
+    public function __construct(string $orderId, ?string $countryId = null, ?string $line1 = null, ?string $line2 = null, ?string $postalCode = null, ?string $city = null, array $data = [])
     {
         $this->orderId = $orderId;
         $this->countryId = $countryId;
@@ -24,6 +32,7 @@ class UpdateBillingAddress
         $this->line2 = $line2;
         $this->postalCode = $postalCode;
         $this->city = $city;
+        $this->data = $data;
     }
 
     public function getOrderId(): OrderId
@@ -40,5 +49,10 @@ class UpdateBillingAddress
             $this->postalCode,
             $this->city,
         );
+    }
+
+    public function getData(): array
+    {
+        return $this->data;
     }
 }