+++ src/Application/Customer/UpdateShippingAddress.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Thinktomorrow\Trader\Application\Customer;
@@ -10,13 +11,20 @@ use Thinktomorrow\Trader\Domain\Model\Customer\CustomerId;
 class UpdateShippingAddress
 {
     private string $customerId;
+
     private ?string $countryId;
+
     private ?string $line1;
+
     private ?string $line2;
+
     private ?string $postalCode;
+
     private ?string $city;
 
-    public function __construct(string $customerId, ?string $countryId = null, ?string $line1 = null, ?string $line2 = null, ?string $postalCode = null, ?string $city = null)
+    private array $data;
+
+    public function __construct(string $customerId, ?string $countryId = null, ?string $line1 = null, ?string $line2 = null, ?string $postalCode = null, ?string $city = null, array $data = [])
     {
         $this->customerId = $customerId;
         $this->countryId = $countryId;
@@ -24,6 +32,7 @@ class UpdateShippingAddress
         $this->line2 = $line2;
         $this->postalCode = $postalCode;
         $this->city = $city;
+        $this->data = $data;
     }
 
     public function getCustomerId(): CustomerId
@@ -40,5 +49,10 @@ class UpdateShippingAddress
             $this->postalCode,
             $this->city,
         );
+    }
+
+    public function getData(): array
+    {
+        return $this->data;
     }
 }